fix: migrate places service to PostgreSQL and resolve import issues
